install: add install package script
It was becoming hard to track what programs I had installed and
how they were configured.
The install scripts are a maintained list of all packages I have
installed, their corresponding config will live in the .dotfiles
directory.

In the event that everything needs to be reinstalled I should
just have to run the install.sh script to install the packages and
then stow-all.sh to create the config files.