xmonad: use spawnon to autostart emacs and firefox
The manageHook would look for any window named firefox or emacs
and move it to their workspace, this meant I could not open
another emacs or firefox session in another workspace because
the manage hook would kick in and move it to the code/web
workspace.

SpawnOn only works at start-up so will not affect any other
instances of emacs or firefox.