polybar: fix clock icon and battery spacing
Upgrading to latest polybar includes some bug-fixes which seem
to have exposed errors in my config.

Date definition had the wrong icon, the char code was not allocated in
font-1 so polybar would fall back to font-3 and use the correct icon
but font-1 has now added an icon for the char code which was being
displayed instead of the clock.

Spaces in the battery label text are now respected which was added
an extra gap between the battery icons,