bin, xmonad: toggle visibility of bar xwindow
To make a window full-screen it is floated and then made full-screen
so that it covers the bar - the bar is not actually removed.

This is an issue for terminal windows which are semi transparent,
because you can still see the bar underneath.

Using xdo will completely remove the bar, but unfortunately it will leave a
gap at the top of all windows where the bar was, you can remove the gap
by toggling structs manually on the workspace(mod-key + n).